requirements.txt
pandas
numpy
requests
pvlib
xgboost
scikit-learn
joblib
matplotlib

# Web / API (added by teammate)
fastapi
uvicorn[standard]
pydantic>=2.0

# Edge Data Layer
paho-mqtt>=1.6.1
sqlalchemy>=2.0.0

# Dev-only: lightweight Python MQTT broker (production use Mosquitto instead)
# amqtt>=0.11.0
